Rename python util
