+++ python/src/vertex.rs
@@ -4,7 +4,7 @@
 use crate::dynamic::DynamicGraph;
 use crate::edge::{PyEdges, PyNestedEdges};
 use crate::types::repr::{iterator_repr, Repr};
-use crate::util::{extract_vertex_ref, through_impl, window_impl};
+use crate::utils::{extract_vertex_ref, through_impl, window_impl};
 use crate::wrappers::iterators::*;
 use crate::wrappers::prop::Prop;
 use docbrown::core::tgraph::VertexRef;
